requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.10.0
gpt4all==2.8.2
greenlet==3.1.1
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.15
langchain-text-splitters==0.3.2
langsmith==0.1.139
lxml==5.3.0
MarkupSafe==3.0.2
minio==7.2.10
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.11
packaging==24.1
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.3.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
yarl==1.17.1
